+++ hooks/post_gen_project.py
@@ -1,0 +1,1 @@
+print("\n\nPost project generation project executed.")