+++ {{cookiecutter.project_slug}}/api_docs/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class APIDocsConfig(AppConfig):
+    name = 'api_docs'