+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for my_project project.
+WSGI config for {{cookiecutter.project_slug}} project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'my_project.config.production')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', '{{cookiecutter.project_slug}}.config.production')
 
 application = get_wsgi_application()